feat: R11 e R12

+++ back-end/src/database/migrations/20230328192217-createProducts.js
@@ -14,7 +14,7 @@ module.exports = {
         allowNull: false,
       },
       price: {
-        type: Sequelize.DOUBLE,
+        type: Sequelize.DECIMAL(4,2),
         allowNull: false,
       },
       urlImage: {
